requirements.txt
flet==0.22.*
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
astor==0.8.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
distro==1.9.0
fastapi==0.114.0
fire==0.6.0
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
fonttools==4.53.1
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
idna==3.8
imageio==2.35.1
imgaug==0.4.0
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==23.2
paddleocr==2.8.1
paddlepaddle==2.6.1
pillow==10.4.0
pip==24.2
proto-plus==1.24.0
protobuf==3.20.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyclipper==1.3.0.post5
pydantic==2.9.1
pydantic-core==2.23.3
Pygments==2.18.0
pyparsing==3.1.4
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
rapidfuzz==3.9.7
repath==0.9.0
requests==2.32.3
rich==13.8.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
setuptools==74.1.2
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.5
termcolor==2.4.0
text-unidecode==1.3
tifffile==2024.8.30
tqdm==4.66.5
types-python-dateutil==2.9.0.20240906
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.0.1